+++ packages/react/src/SignUpForm/NewUserSignUp.stories.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { Meta } from '@storybook/react';
 
 import { NileProvider } from '../context';
+import theme from '../../.storybook/themeConfig';
 
 import SignUpForm from '.';
 
@@ -14,8 +15,11 @@ export default meta;
 
 export function Basic() {
   return (
-    <NileProvider>
-      <div style={{ maxWidth: '20rem', margin: '0 auto' }}>
+    <NileProvider {...theme}>
+      <div
+        style={{ maxWidth: '20rem', margin: '0 auto' }}
+        data-joy-color-scheme="dark"
+      >
         <SignUpForm onSuccess={() => alert('success!')} />
       </div>
     </NileProvider>