+++ packages/server/src/index.ts
@@ -2,3 +2,5 @@ export * from './types';
 import { default as Server } from './Server';
 
 export default Server;
+
+module.exports = Server;